pyproject.toml
[tool.poetry]
name = "stv-services"
version = "0.1.0"
description = "Integration services for Seed The Vote"
authors = ["Daniel Brotsky <dan@clickonetwo.io>"]

[tool.poetry.dependencies]
python = "^3.9"
SQLAlchemy = {version = "^1.4.32", extras = ["postgresql"]}
alembic = {extras = ["tz"], version = "^1.7.7"}
python-dotenv = "^0.20.0"
restnavigator = "^1.0.1"
python-dateutil = "^2.8.2"
click = "^8.1.2"
pyairtable = "^1.1.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pytest = "^7.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
